64574: Further refactoring of item-component, mantis publication and JSDocs
