+++ themes/mantis/app/+item-page/simple/item-types/publication/publication.component.html
@@ -62,19 +62,19 @@
         <div class="row">
             <div class="col-12 col-md-4" *ngIf="projects$ | async">
                 <ds-related-items
-                        [items]="projects$ | async"
+                        [items]="(projects$ | async)?.payload?.page"
                         [label]="'relationships.isProjectOf' | translate">
                 </ds-related-items>
             </div>
             <div class="col-12 col-md-4" *ngIf="orgUnits$ | async">
                 <ds-related-items
-                        [items]="orgUnits$ | async"
+                        [items]="(orgUnits$ | async)?.payload?.page"
                         [label]="'relationships.isOrgUnitOf' | translate">
                 </ds-related-items>
             </div>
             <div class="col-12 col-md-4" *ngIf="journalIssues$ | async">
                 <ds-related-items
-                        [items]="journalIssues$ | async"
+                        [items]="(journalIssues$ | async)?.payload?.page"
                         [label]="'relationships.isJournalIssueOf' | translate">
                 </ds-related-items>
             </div>
